+++ cmd/bounce/handlers.go
@@ -9,6 +9,7 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -40,15 +41,29 @@ const (
 // TODO: Find a better location for this constant
 const MB int64 = 1048576
 
+// clientRequestDetails is used to bundle various client request details for
+// processing by templates or notification functions.
+type clientRequestDetails struct {
+	Datestamp          string
+	EndpointPath       string
+	HTTPMethod         string
+	ClientIPAddress    string
+	Headers            http.Header
+	Body               string
+	BodyError          string
+	FormattedBody      string
+	FormattedBodyError string
+	RequestError       string
+	ContentTypeError   string
+}
+
 // handleIndex receives our HTML template and our defined routes as a pointer.
 // Both are used to generate a dynamic index of the available routes or
 // "endpoints" for users to target with test payloads. A pointer is used because
 // by the time this handler is defined, the full set of routes has *not* been
 // defined. Using a pointer, we are able to access the complete collection
 // of defined routes when this handler is finally called.
-func handleIndex(templateText string, rs *routes.Routes) http.HandlerFunc {
-
-	tmpl := htmlTemplate.Must(htmlTemplate.New("indexPage").Parse(templateText))
+func handleIndex(tmpl *htmlTemplate.Template, rs *routes.Routes) http.HandlerFunc {
 
 	return func(w http.ResponseWriter, r *http.Request) {
 
@@ -65,7 +80,7 @@ func handleIndex(templateText string, rs *routes.Routes) http.HandlerFunc {
 				"http_method": r.Method,
 			}).Debug("non-GET request received on GET-only endpoint")
 			errorMsg := fmt.Sprintf(
-				"\nSorry, this endpoint only accepts %s requests.\n"+
+				"Sorry, this endpoint only accepts %s requests. "+
 					"Please see the README for examples and then try again.",
 				http.MethodGet,
 			)
@@ -91,6 +106,7 @@ func handleIndex(templateText string, rs *routes.Routes) http.HandlerFunc {
 		err := tmpl.Execute(w, *rs)
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
+			ctxLog.Error(err.Error())
 		}
 
 	}
@@ -98,48 +114,49 @@ func handleIndex(templateText string, rs *routes.Routes) http.HandlerFunc {
 }
 
 // echoHandler echos back the HTTP request received by
-func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) http.HandlerFunc {
+func echoHandler(ctx context.Context, tmpl *textTemplate.Template, coloredJSON bool, coloredJSONIndent int, notifyWorkQueue chan<- clientRequestDetails) http.HandlerFunc {
 
 	return func(w http.ResponseWriter, r *http.Request) {
 
 		// For now, we generate plain text responses
 		//w.Header().Set("Content-Type", "text/plain")
 
-		type echoHandlerResponse struct {
-			Datestamp          string
-			EndpointPath       string
-			HTTPMethod         string
-			ClientIPAddress    string
-			Headers            http.Header
-			Body               string
-			BodyError          string
-			FormattedBody      string
-			FormattedBodyError string
-			RequestError       string
-			ContentTypeError   string
-		}
-
-		ourResponse := echoHandlerResponse{}
+		ourResponse := clientRequestDetails{}
 
 		mw := io.MultiWriter(w, os.Stdout)
 
-		tmpl := textTemplate.Must(textTemplate.New("echoHandler").Parse(templateText))
-
+		// TODO: Consider moving this "up" so that it can receive values as
+		// arguments instead of relying on them to be defined in the local
+		// scope?
 		writeTemplate := func() {
 			err := tmpl.Execute(mw, ourResponse)
 			if err != nil {
 				http.Error(w, err.Error(), http.StatusInternalServerError)
+				log.Errorf("error occurred while trying to execute template: %v", err)
 
 				// We force a a return here since it is unlikely that we should
 				// execute any other code after failing to generate/write out our
 				// template
 				return
 			}
+
+			// Manually flush http.ResponseWriter
+			// https://blog.simon-frey.eu/manual-flush-golang-http-responsewriter/
+			if f, ok := w.(http.Flusher); ok {
+				log.Debug("echoHandler: Manually flushing http.ResponseWriter")
+				f.Flush()
+			} else {
+				log.Warn("echoHandler: http.Flusher interface not available, cannot flush http.ResponseWriter")
+				log.Warn("echoHandler: Not flushing http.ResponseWriter may cause a noticeable delay between requests")
+			}
+
 		}
 
-		log.Debug("echoHandler endpoint hit")
+		log.Debug("echoHandler: echoHandler endpoint hit")
 
-		ourResponse.Datestamp = time.Now().Format((time.RFC3339))
+		// Work around Teams choosing to ignore time.RFC3339 designation and
+		// display as localtime by explicitly converting to localtime
+		ourResponse.Datestamp = time.Now().Format("2006-01-02 15:04:05")
 		ourResponse.EndpointPath = r.URL.Path
 		ourResponse.HTTPMethod = r.Method
 		ourResponse.ClientIPAddress = GetIP(r)
@@ -154,8 +171,12 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 
 			case http.MethodGet:
 
-				// Write out what we have. Nothing further to do for this endpoint
+				// Write out what we have.
 				writeTemplate()
+
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
 				return
 
 			case http.MethodPost:
@@ -168,7 +189,13 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 					ourResponse.BodyError = errorMsg
 
 					http.Error(w, errorMsg, http.StatusBadRequest)
+					log.Error(errorMsg)
+
 					writeTemplate()
+
+					// Send to Notification Manager for further processing
+					go func() { notifyWorkQueue <- ourResponse }()
+
 					return
 				}
 
@@ -184,12 +211,23 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 				// the template against it
 				writeTemplate()
 
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
+				return
+
 			default:
 				errorMsg := fmt.Sprintf("ERROR: Unsupported method %q received; please try again using %s method", r.Method, http.MethodPost)
 				ourResponse.RequestError = errorMsg
 
 				http.Error(w, errorMsg, http.StatusMethodNotAllowed)
+				log.Error("echoHandler: " + errorMsg)
+
 				writeTemplate()
+
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
 				return
 			}
 
@@ -200,14 +238,20 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 			case http.MethodGet:
 				// TODO: Collect this for use with our template
 				errorMsg := fmt.Sprintf(
-					"Sorry, this endpoint only accepts JSON data via %s requests.\n"+
+					"Sorry, this endpoint only accepts JSON data via %s requests. "+
 						"Please see the README for examples and then try again.",
 					http.MethodPost,
 				)
 				ourResponse.RequestError = errorMsg
 
 				http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+				log.Error("echoHandler: " + errorMsg)
+
 				writeTemplate()
+
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
 				return
 
 			case http.MethodPost:
@@ -227,7 +271,13 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 					ourResponse.BodyError = errorMsg
 
 					http.Error(w, errorMsg, http.StatusBadRequest)
+					log.Error(errorMsg)
+
 					writeTemplate()
+
+					// Send to Notification Manager for further processing
+					go func() { notifyWorkQueue <- ourResponse }()
+
 					return
 				}
 				ourResponse.Body = string(requestBody)
@@ -246,15 +296,25 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 							ourResponse.FormattedBodyError = fmt.Sprintf("%s: %s", errorPrefix, mr.msg)
 
 							http.Error(w, mr.msg, mr.status)
+
 							writeTemplate()
+
+							// Send to Notification Manager for further processing
+							go func() { notifyWorkQueue <- ourResponse }()
+
 							return
 						}
 
 						errorMsg := fmt.Sprintf("%s: %s", errorPrefix, err.Error())
 						ourResponse.FormattedBodyError = errorMsg
-						log.Error(errorMsg)
 						http.Error(w, errorMsg, http.StatusInternalServerError)
+						log.Error("echoHandler: " + errorMsg)
+
 						writeTemplate()
+
+						// Send to Notification Manager for further processing
+						go func() { notifyWorkQueue <- ourResponse }()
+
 						return
 					}
 				}
@@ -294,20 +354,32 @@ func echoHandler(templateText string, coloredJSON bool, coloredJSONIndent int) h
 				// the template against it
 				writeTemplate()
 
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
 			default:
 				errorMsg := fmt.Sprintf("ERROR: Unsupported method %q received; please try again using %s method", r.Method, http.MethodPost)
 				ourResponse.RequestError = errorMsg
 
 				http.Error(w, errorMsg, http.StatusMethodNotAllowed)
+
 				writeTemplate()
+
+				// Send to Notification Manager for further processing
+				go func() { notifyWorkQueue <- ourResponse }()
+
 				return
 			}
 
 		default:
 			// Template is not used for this code block, so no need to account for
 			// the output in the template
-			log.Debugf("Rejecting request %q; not explicitly handled by a route.", r.URL.Path)
+			log.Debugf("echoHandler: Rejecting request %q; not explicitly handled by a route.", r.URL.Path)
 			http.NotFound(w, r)
+
+			// Send to Notification Manager for further processing
+			go func() { notifyWorkQueue <- ourResponse }()
+
 			return
 		}